requirements.txt
ultralytics
torch
torchvision
transformers
timm
numpy
opencv-python
pillow